--- a/schema.go
+++ b/schema.go
@@ -932,12 +932,6 @@ func createColumnDefinitionFromColumn(c *Column) *parquetschema.ColumnDefinition
 
 func (r *schema) GetSchemaDefinition() *parquetschema.SchemaDefinition {
 	return r.schemaDef
-	//def, err := parquetschema.ParseSchemaDefinition(r.schemaDef.String())
-	//if err != nil {
-	//	panic(err)
-	//}
-	//
-	//return def
 }
 
 // DataSize return the size of data stored in the schema right now