## BEFORE - 27511/s
```
BenchmarkParquetConsumer-12        33078             36349 ns/op            5617 B/op        162 allocs/op
PASS
ok      github.com/clearbit/stream-utils/partitioner/writers/parquet    1.727s
```



## After - 110521/s
```
BenchmarkParquetConsumer-12       111354              9048 ns/op            2387 B/op         80 allocs/op
PASS
ok      github.com/clearbit/stream-utils/partitioner/writers/parquet    1.300s
```